front end updated and item and customer forms are completed

+++ POS_BackEnd/src/main/java/lk/ijse/spring/dto/ItemDTO.java
@@ -6,8 +6,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
 
-import java.math.BigDecimal;
-
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
@@ -16,6 +14,6 @@ public class ItemDTO {
     private String code;
     private String description;
     private int qtyOnHand;
-    private BigDecimal unitPrice;
+    private double unitPrice;
 
 }
